# server/src/main/resources/api/openapi.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Tesseract
  description: An API for tesseract
  termsOfService: https://swagger.io/terms/
  contact:
    name: Swagger API Team
    email: apiteam@swagger.io
    url: https://swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /diversifications:
    get:
      tags:
        - diversification
      summary: Получить список диверсификаций
      parameters:
        - $ref: '#/components/parameters/authCredsParam'
        - $ref: '#/components/parameters/pageNumber'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: Успешное получение списка диверсификаций
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiversificationsList'
        '401':
          description: Отсутствуют данные аутентификации
    post:
      tags:
        - diversification
      summary: Создать диверсификацию
      requestBody:
        description: Данные для создании диверсификации
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiversificationCreateInfo'
      parameters:
        - $ref: '#/components/parameters/authCredsParam'
      responses:
        '201':
          description: Успешное создание диверсификации
        '400':
          description: Некорректное тело запроса
        '401':
          description: Отсутствуют данные аутентификации
        '404':
          description: Отсутствуют активы с выбранным типом рискованности
  /diversifications/{id}:
    get:
      tags:
        - diversification

      summary: Получить информацию о конкретной диверсификации
      parameters:
        - $ref: '#/components/parameters/authCredsParam'
        - $ref: '#/components/parameters/idPath'
      responses:
        '200':
          description: Успешное получение диверсификации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecifiedDiversification'
        '400':
          description: Некорректное тело запроса
        '401':
          description: Отсутствуют данные аутентификации
        '404':
          description: Отсутствует диверсификация с таким id (либо запрашиваем несвою диверсификацию)
components:
  parameters:
    authCredsParam:
      name: authCreds
      in: header
      description: Данные аутентификации
      required: true
      schema:
        type: string
    idPath:
      name: id
      in: path
      description: Уникальный номер сущности
      required: true
      schema:
        type: integer
    pageNumber:
      name: pageNumber
      in: query
      description: Номер страницы пагинации (начинается с 0)
      required: false
      schema:
        type: integer
    pageSize:
      name: pageSize
      in: query
      description: Размер одной страницы
      required: false
      schema:
        type: integer
  schemas:
...
    Diversification:
      type: object
      properties:
        diversificationId:
          type: integer
        createDatetime:
          type: string
          description: Формат времени ISO 8601
        riskTypeId:
          type: integer
        amount:
          type: integer
          description: Сумма диверсификации, в копейках
    DiversificationsList:
      type: array
      items:
        $ref: '#/components/schemas/Diversification'
    DiversificationCreateInfo:
      type: object
      properties:
        amount:
          type: integer
          description: Сумма создаваемой диверсификации, в копейках
        riskTypeId:
          type: integer
    DiversificationAsset:
      type: object
      properties:
        assetId:
          type: integer
        assetTitle:
          type: string
        companyTitle:
          type: string
        oldPrice:
          type: integer
          description: Старая цена актива (то есть цена на момент создания диверсификации), в копейках 
        count:
          type: integer
    SpecifiedDiversification:
      type: object
      properties:
        createDatetime:
          type: string
          description: Формат времени ISO 8601
        amount:
          type: integer
        riskTypeId:
          type: integer
        assetsList:
          type: array
          items:
            $ref: "#/components/schemas/DiversificationAsset"
...